requirements/dev.txt
-r base.txt
django-debug-toolbar
selenium
mysqlclient
sqlparse
